# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "Guardian"
  description: "The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation."
  version: "2.9.0"
  contact:
    name: "API developer"
    url: "https://envisionblockchain.com"
    email: "info@envisionblockchain.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
- url: /api/v1
  description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
...
    User:
      type: object
      required:
        - username
        - role
      properties:
        confirmed:
          type: string
        failed:
          type: string
        username:
          type: string
        role:
          type: string
        hederaAccountId:
          type: string
        hederaAccountKey:
          type: string
        did:
          type: string
        didDocument:
            type: object
        vcDocument:
            type: object
        parent:
          type: string
        topicId:
          type: string
...
    TrustChains:
      type: object
      required:
        - chain
        - userMap
      properties:
        chain:
          type: array
          items:
            type: object
            required:
              - id
              - type
              - tag
              - label
              - schema
              - owner
              - document
            properties:
              id:
                type: string
              type:
                type: string
              tag:
                type: string
              label:
                type: string
              schema:
                type: string
              owner:
                type: string
              document:
                type: object
        userMap:
          type: array
          items:
            type: object
            required:
              - did
              - username
            properties:
              did:
                type: string
              username:
                type: string
    VerifiablePresentation:
      type: object
      required:
        - hash
        - id
        - policyId
        - signature
        - status
        - tag
        - type
        - updateDate
        - createDate
        - owner
        - document
      properties:
        hash:
          type: string
        id:
          type: string
        policyId:
          type: string
        signature:
          type: string
        status:
          type: string
        tag:
          type: string
        type:
          type: string
        updateDate:
          type: string
        createDate:
          type: string
        owner:
          type: string
        document:
          type: object
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    CommonSettings:
      type: object
      properties:
        operatorId:
          type: string
        operatorKey:
          type: string
        nftApiKey:
          deprecated: true
          type: string
        ipfsStorageApiKey:
          type: string
...
    Task:
      type: object
      properties:
        taskId:
          type: string
        expectation:
          type: number
...
paths:
...
  /profiles/{username}/balance:
    get:
      tags:
        - profiles
      description: Requests Hedera account balance. Only users with the Installer role are allowed to make the request.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to fetch the balance.
      security:
      - bearerAuth: []
      summary: Returns user's Hedera account balance.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: string
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profiles/{username}:
    get:
      tags:
        - profiles
      description: Returns user account information. For users with the Standard Registry role it also returns address book and VC document information.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to fetch the information.
      security:
      - bearerAuth: []
      summary: Returns user account info.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - profiles
      description: Sets Hedera credentials for the user. For users with the Standard Registry role it also creates an address book.
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to update the information.
      security:
      - bearerAuth: []
      summary: Sets Hedera credentials for the user.
      requestBody:
        description: Object that contains the Hedera account data.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        201:
          description: Successful operation.
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profiles/push/{username}:
    put:
      tags:
        - profiles
      description: Sets Hedera credentials for the user
      security:
        - bearerAuth: []
      summary: Sets Hedera credentials for the user
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user for whom to update the information.
      requestBody:
        description: Object that contains the Hedera account data.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /trustchains:
    get:
      tags:
        - trustchains
      description: Requests all VP documents. Only users with the Auditor role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Returns a list of all VP documents.
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: The number of pages to skip before starting to collect the result set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
        - in: query
          name: policyId
          schema:
            type: string
          description: Selected policy ID.
        - in: query
          name: policyOwner
          schema:
            type: string
          description: Selected Standard Registry (DID).
      responses:
        200:
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VerifiablePresentation'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /trustchains/{hash}:
    get:
      tags:
        - trustchains
      description: Builds and returns a trustchain, from the VP to the root VC document. Only users with the Auditor role are allowed to make the request.
      parameters:
        - in: path
          name: hash
          schema:
            type: string
          required: true
          description: Hash or ID of a VP document.
      security:
      - bearerAuth: []
      summary: Returns a trustchain for a VP document.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustChains'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /ipfs/file/{cid}:
    get:
      tags:
        - ipfs
      description: Get file from ipfs.
      summary: Get file from ipfs.
      parameters:
        - in: path
          name: cid
          schema:
            type: string
          required: true
          description: File CID.
      security:
      - bearerAuth: []
      responses:
        201:
          description: Successful operation.
          content:
            binary/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /settings:
    get:
      tags:
        - settings
      description: Returns current settings. For users with the Standard Registry role only.
      security:
      - bearerAuth: []
      summary: Returns current settings.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonSettings'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - settings
      description: Set settings. For users with the Standard Registry role only.
      security:
      - bearerAuth: []
      summary: Set settings.
      responses:
        200:
          description: Successful operation.
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonSettings'
  /settings/environment:
    get:
      tags:
        - settings
      description: Returns current environment name.
      security:
      - bearerAuth: []
      summary: Returns current environment name.
      responses:
        200:
          description: Successful operation.
          content:
            text/plain:
              schema:
                type: string
                example: 'testnet'
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...